requirements.txt
inputs
numpy
pillow
opencv-python
torch===1.7.0
torchvision===0.8.1
torchaudio===0.7.0
fastai